Safari-611.1.10/LayoutTests/http/tests/resourceLoadStatistics/downgrade-document-referrer-third-party-iframe-expected.txt

Tests that document.referrer is downgraded in third-party iframes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.referrer is http://127.0.0.1:8000
PASS successfullyParsed is true

TEST COMPLETE
